requirements.txt
streamlit==1.44.0
streamlit_code_editor==0.1.22
